+++ WTMGlyph/WTMGlyphUtilities.h
@@ -12,14 +12,14 @@
 
 typedef struct FloatArrayContainer {
     float * items;
-    int itemCount;
-    int allocatedCount;
+    NSUInteger itemCount;
+    NSUInteger allocatedCount;
 } FloatArrayContainer;
 
 NSArray* Resample(NSArray *points, int num);
 NSArray* Scale(NSArray *points, int resolution, float threshold);
 CGRect BoundingBox(NSArray *points);
-NSMutableArray* Splice(NSMutableArray *original, id newVal, int i);
+NSMutableArray* Splice(NSMutableArray *original, id newVal, NSUInteger i);
 float PathLength(NSArray *points);
 float Distance(CGPoint point1, CGPoint point2);
 float IndicativeAngle(NSArray *points);
